+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+### next release
+- :bulb: Extended `Users::Queries::Attributes` query to accept `users` array
+
 ### v0.3.1
 - :bulb: Added `ActiveFriends` users query
 - :hammer: Refactored `Friends` users query